requirements.txt
#pep8==1.7.1
#autopep8==2.0.1
#fiona==1.9.1
#geopandas==0.12.2
pillow==9.4.0
selenium==4.13.0
requests==2.28.1
chardet==5.1.0
